This upgrades the initialization of both camera's of the Bebop. It also includes an ISP configuration to convert front camera images to UYVY.

+++ sw/airborne/modules/computer_vision/cv.h
@@ -55,7 +55,7 @@ struct video_listener {
     cv_function func;
 
     // Can be set by user
-    int maximum_fps;
+    uint16_t maximum_fps;
     volatile bool active;
 };
